added casting to test classes because im silly

+++ src/main/resources/classundertest/AByteADay.java
@@ -3,11 +3,11 @@ package com.github.bradleyrumball.autologic;
 public class AByteADay {
 
     public static String howCrunchyIsMyApple(byte num) {
-        if (num == 6) {
+        if (num == (byte)6) {
             return "Strong cronch";
-        } else if (num > 10) {
+        } else if (num > (byte)10) {
             return "Large Apple";
-        } else if (num < 2) {
+        } else if (num < (byte)2) {
             return "Small crunch";
         } else {
             return "No crunch";
