+++ src/main/resources/classundertest/FloatAway.java
@@ -3,7 +3,7 @@ package com.github.bradleyrumball.autologic;
 public class FloatAway {
 
     public static String wheresBoatNumber(float num) {
-        if (num == -6.5) {
+        if (num == (float)-6.5) {
             return "Floated away mate";
         } else {
             return "No boat";
