+++ src/main/resources/classundertest/LongTests.java
@@ -3,9 +3,9 @@ package com.github.bradleyrumball.autologic;
 public class LongTests {
 
     public static String LomgTests(long num) {
-        if (num > 60000) {
+        if (num > (long)60000) {
             return "Very long";
-        } else if (num < 2) {
+        } else if (num < (long)2) {
             return "Not very long";
         } else {
             return "Nothing to see here";
